Create GUI with querying info and current problems
here is the general code of everything we need just to get the ball rolling. the extra pop up menu can be added after i clear the part of having the proper information fetched based off of what the user inputs and which filter is selected. this is still based off my example database which i will upload, if you want to use that to run the code